refactor choose ai args into 2 files
